+++ src/GridPartition.h
@@ -33,6 +33,7 @@ class GridPartition
     template<typename T> void lookup(T& shape, std::vector<Point*>&);
     template<typename T> void lookup(T& shape, std::vector<PointXYZ>&);
     void knn(const Point&, const unsigned int, std::vector<Point*>&);
+    void knn(const Point&, const unsigned int, const double, std::vector<Point*>&);
     void knn(const PointXYZ&, const unsigned int, std::vector<PointXYZ>&);
     void knn(const PointXYZ&, const unsigned int, const double, std::vector<PointXYZ>&);