+++ NEWS.md
@@ -114,6 +114,8 @@ In attempt to do not break users scripts the version 3 is fully backward compati
 
 17. New function `get_range()`.
 
+18. `knnidw()` gains an argument `rmax` to set a maximum radius search where to find the knn. This fixes computation time issues with non convex point clouds.
+
 #### ENHANCEMENT
 
 1. `readLAS()` now warns when reading incompatible files. Point coordinates are recomputed on-the-fly as it has always been done but now the user is aware of potential trouble or precision loss.